% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentiment_plot.R
\name{get_sentiment_plot}
\alias{get_sentiment_plot}
\title{Brings out the sentiment report for the most common words from the top news search results}
\usage{
get_sentiment_plot(company_name)
}
\arguments{
\item{company_name}{name of the company}
}
\value{
list
}
\description{
Brings out the sentiment report for the most common words from the top news search results
}
\examples{
get_sentiment_plot("Debenhams")
get_sentiment_plot("Vodafone")
}
