% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wordcloud.R
\name{get_wordcloud}
\alias{get_wordcloud}
\title{Brings out the wordclouds for the most common words from the top news search results}
\usage{
get_wordcloud(company_name)
}
\arguments{
\item{company_name}{name of the company}
}
\value{
list
}
\description{
Brings out the wordclouds for the most common words from the top news search results
}
\examples{
get_wordcloud("Debenhams")
get_wordcloud("Vodafone")
}
