% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_words.R
\name{get_words}
\alias{get_words}
\title{Brings out the most common words from the top news search results}
\usage{
get_words(company_name)
}
\arguments{
\item{company_name}{name of the company}
}
\value{
list
}
\description{
Brings out the most common words from the top news search results
}
\examples{
get_words("Debenhams")
get_words("Vodafone")
}
