% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_words_tfidf.R
\name{get_words_tfidf}
\alias{get_words_tfidf}
\title{Brings out the most common words from the top news search results with tf-idf}
\usage{
get_words_tfidf(company_name)
}
\arguments{
\item{company_name}{name of the company}
}
\value{
list
}
\description{
Brings out the most common words from the top news search results with tf-idf
}
\examples{
get_words_tfidf("Debenhams")
get_words_tfidf("Vodafone")
}
