% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_article.R
\name{which_article}
\alias{which_article}
\title{Brings out the must read articles from the top news search results}
\usage{
which_article(company_name)
}
\arguments{
\item{company_name}{name of the company}
}
\value{
character
}
\description{
Brings out the must read articles from the top news search results
}
\examples{
which_article("Debenhams")
which_article("Vodafone")
}
